typed options && restructure ChartsDrawer
